requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
redis==5.0.1
pinecone-client==3.0.0
openai==1.10.0
rank-bm25==0.2.2
numpy==1.26.3
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.0
pyjwt==2.8.0
pypdf==3.17.4
tiktoken==0.5.2
python-magic==0.4.27
prometheus-client==0.19.0
httpx==0.27.0
langchain-text-splitters==0.3.8

python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.2.0
beautifulsoup4==4.12.3
lxml==5.2.1
ebooklib==0.20
pymupdf==1.23.26
faster-whisper==1.0.3
pillow==10.3.0
paddleocr
paddlepaddle
opencv-python-headless
pdf2image
celery==5.4.0
pydantic-settings==2.2.1
